crackapi

+++ packages/api_asscrack/src/crack_worker/mod.rs
@@ -7,7 +7,7 @@ pub struct WorkerPipe {
 
 #[derive(Clone, serde::Serialize, serde::Deserialize, Debug)]
 pub struct WorkerMessage {
-    pub msg_id: u64,
+    pub msg_id: u32,
     pub msg_type: String,
     pub msg_content: Vec<u8>,
 }
